+++ demo/src/demos/token/demo.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { RouterModule } from '@angular/router';
-import { VCLTokenModule} from '@ng-vcl/ng-vcl';
+import { VCLTokenModule, VCLInputModule } from '@ng-vcl/ng-vcl';
 import { DemoModule, DemoComponent } from './../../modules/demo/demo.module';
 import { TokenDemoComponent } from './demo.component';
 
@@ -33,6 +33,7 @@ export function demo() {
     FormsModule,
     DemoModule,
     VCLTokenModule,
+    VCLInputModule,
     RouterModule.forChild([{
       path: '',
       component: DemoComponent,