+++ src/ng-vcl/input/index.ts
@@ -1,12 +1,15 @@
+import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
-import { InputDirective }   from './input.directive';
+import { VCLWormholeModule } from './../wormhole/index';
+import { VCLPopoverModule } from '../popover/index';
+import { InputDirective, InputAutoSelectDirective }   from './input.directive';
 
 export { InputDirective };
 
 @NgModule({
   imports: [],
-  exports: [InputDirective],
-  declarations: [InputDirective],
+  exports: [InputDirective, InputAutoSelectDirective],
+  declarations: [InputDirective, InputAutoSelectDirective],
   providers: [],
 })
 export class VCLInputModule { }