+++ src/ng-vcl/password-input/index.ts
@@ -3,13 +3,14 @@ import { FormsModule } from '@angular/forms';
 import { CommonModule } from '@angular/common';
 import { VCLInputModule } from '../input/index';
 import { VCLButtonModule } from '../button/index';
-import { PasswordInputComponent }   from './password-input.component';
+import { PasswordInputContainerComponent }   from './password-input-container.component';
+import { PasswordInputDirective }   from './password-input.directive';
 
-export { PasswordInputComponent };
+export { PasswordInputContainerComponent, PasswordInputDirective };
 
 @NgModule({
   imports: [ FormsModule, VCLInputModule, VCLButtonModule ],
-  exports: [PasswordInputComponent],
-  declarations: [PasswordInputComponent]
+  exports: [PasswordInputContainerComponent, PasswordInputDirective],
+  declarations: [PasswordInputContainerComponent, PasswordInputDirective]
 })
 export class VCLPasswordInputModule { }