+++ src/ng-vcl/token/token.component.ts
@@ -14,6 +14,7 @@ export interface Token {
   label: string;
   selected?: boolean;
   removable?: boolean;
+  value?: any;
 }
 
 @Component({
@@ -31,6 +32,9 @@ export class TokenComponent implements Token {
 
   @Input()
   label: string;
+
+  @Input()
+  value: string;
 
   @Input()
   disabled: boolean = false;